Refactor front-end
