+++ server/controllers/BrazilStatesController.js
@@ -1,7 +1,10 @@
 import api from '../services/ApiIbgeStates';
 
 export default async function BrazilStatesController(req, res) {
-  const { data } = await api.get();
-
-  return res.json(data);
+  try {
+    const { data } = await api.get();
+    return res.json(data);
+  } catch (err) {
+    return res.status(400).json({ error: err });
+  }
 }
