+++ server/controllers/WallpaperController.js
@@ -1,9 +1,11 @@
 import api from '../services/ApiWallpaper';
 
 export default async function WallpaperController(req, res) {
-  const { data } = await api.get();
-
-  const urlWallpaper = `url('https://www.bing.com${data.images[0].url}')`;
-
-  return res.json(urlWallpaper);
+  try {
+    const { data } = await api.get();
+    const urlWallpaper = `url('https://www.bing.com${data.images[0].url}')`;
+    return res.json(urlWallpaper);
+  } catch (err) {
+    return res.status(400).json({ error: err });
+  }
 }
